src/requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-allauth==0.39.1
django-cors-headers==3.0.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
Pillow==6.1.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
stripe==2.32.0
urllib3==1.25.3
whitenoise==4.1.2
